pyproject.toml
[project]
name = "fufx"
version = "0.1.0"
description = "GovInfo daily document crawler with Supabase sync"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
fufx-crawl = "crawler.govinfo:main"
fufx-sync = "supabase.sync:main"

